# Conflicts:
#   Backend/src/main/java/com/filesharing/iot/controller/PeerController.java
#   Backend/src/main/java/com/filesharing/iot/controller/UserController.java
#   Frontend/src/App.js
#   Frontend/src/components/utils/Authorization.js

+++ DesktopApp/mavencharlie/src/main/java/com/iot/desktop/models/FileMetadata.java
@@ -12,7 +12,7 @@ public class FileMetadata {
 
     public FileMetadata() {}
 
-    public FileMetadata(String id, String userId, String groupId, String fileName, String extension, int size, String MD5Signature) {
+    public FileMetadata(String id, String userId, String groupId, String fileName, String extension, long size, String MD5Signature) {
         this.id = id;
         this.userId = userId;
         this.groupId = groupId;
